hobby_loader/requirements.txt
beautifulsoup4 
brotlipy 
certifi==2021.10.8
cffi 
chardet==3.0.4
click 
cryptography 
idna 
neo4j 
pycparser 
pyOpenSSL 
PySocks 
python-dotenv
pytz 
requests 
six==1.15.0
soupsieve==2.0.1
urllib3 
